+++ crates/ruma-events/src/rtc.rs
@@ -1,0 +1,4 @@
+//! Modules for events in the `m.rtc` namespace.
+
+#[cfg(feature = "unstable-msc4310")]
+pub mod decline;